backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.8.0
supabase>=2.13.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9